# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
  description: MANTHANA-SCHOLER — AI-powered thesis and research writing platform for Indian medical scholars
servers:
  - url: /api
paths:
...
  /workspaces/{workspaceId}/sections:
    get:
      operationId: listSections
      summary: List sections in a workspace
      tags: [sections]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"

    post:
      operationId: createSection
      summary: Create a section
      tags: [sections]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
  /workspaces/{workspaceId}/sections/{sectionId}:
    get:
      operationId: getSection
      summary: Get a section
      tags: [sections]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: sectionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"

    patch:
      operationId: updateSection
      summary: Update a section
      tags: [sections]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: sectionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"

    delete:
      operationId: deleteSection
      summary: Delete a section
      tags: [sections]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: sectionId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workspaces/{workspaceId}/sections/reorder:
    post:
      operationId: reorderSections
      summary: Reorder all sections in a workspace
      tags: [sections]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionReorderInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"

  # ─── AI CHAT ─────────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/vault:
    get:
      operationId: listVaultResources
      summary: List vault resources for a workspace
      tags: [vault]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultResource"

    post:
      operationId: createVaultResource
      summary: Add a resource to the vault
      tags: [vault]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultResourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultResource"
  /workspaces/{workspaceId}/vault/summary:
    get:
      operationId: getVaultSummary
      summary: Vault stats by type
      tags: [vault]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultSummary"
  /workspaces/{workspaceId}/vault/{resourceId}:
    get:
      operationId: getVaultResource
      summary: Get a vault resource
      tags: [vault]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: resourceId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultResource"

    patch:
      operationId: updateVaultResource
      summary: Update a vault resource
      tags: [vault]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: resourceId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultResourceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultResource"

    delete:
      operationId: deleteVaultResource
      summary: Delete a vault resource
      tags: [vault]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: resourceId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Section:
      type: object
      required: [id, workspaceId, title, type, status, order, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [introduction, literature_review, methodology, results, discussion, conclusion, abstract, references, custom]
        content:
          type: ["string", "null"]
        status:
          type: string
          enum: [not_started, in_progress, completed, reviewed]
        order:
          type: integer
        wordCount:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SectionInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [introduction, literature_review, methodology, results, discussion, conclusion, abstract, references, custom]
        content:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, completed, reviewed]
        order:
          type: integer
        notes:
          type: string
    SectionUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [introduction, literature_review, methodology, results, discussion, conclusion, abstract, references, custom]
        content:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, completed, reviewed]
        order:
          type: integer
        notes:
          type: string
    SectionReorderInput:
      type: object
      required: [orderedIds]
      properties:
        orderedIds:
          type: array
          items:
            type: integer

    # ─── CHAT ────────────────────────────────────────────────────────────────
...
    VaultResource:
      type: object
      required: [id, workspaceId, type, title, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        type:
          type: string
          enum: [paper, note, reference, image, link]
        title:
          type: string
        content:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        authors:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        journal:
          type: ["string", "null"]
        doi:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    VaultResourceInput:
      type: object
      required: [type, title]
      properties:
        type:
          type: string
          enum: [paper, note, reference, image, link]
        title:
          type: string
          minLength: 1
        content:
          type: string
        url:
          type: string
        authors:
          type: string
        year:
          type: integer
        journal:
          type: string
        doi:
          type: string
        tags:
          type: string
    VaultResourceUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        url:
          type: string
        authors:
          type: string
        year:
          type: integer
        journal:
          type: string
        doi:
          type: string
        tags:
          type: string
    VaultSummary:
      type: object
      required: [total, byType, recentResources]
      properties:
        total:
          type: integer
        byType:
          type: object
          properties:
            paper:
              type: integer
            note:
              type: integer
            reference:
              type: integer
            image:
              type: integer
            link:
              type: integer
        recentResources:
          type: array
          items:
            $ref: "#/components/schemas/VaultResource"