# lib/api-zod/src/generated/types/vaultResource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MANTHANA-SCHOLER — AI-powered thesis and research writing platform for Indian medical scholars
 * OpenAPI spec version: 1.0.0
 */
import type { VaultResourceType } from './vaultResourceType';

export interface VaultResource {
  id: number;
  workspaceId: number;
  type: VaultResourceType;
  title: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  authors?: string | null;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  journal?: string | null;
  /** @nullable */
  doi?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
  updatedAt: string;
}