# lib/api-zod/src/generated/types/vaultResourceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MANTHANA-SCHOLER — AI-powered thesis and research writing platform for Indian medical scholars
 * OpenAPI spec version: 1.0.0
 */

export interface VaultResourceUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  url?: string;
  authors?: string;
  year?: number;
  journal?: string;
  doi?: string;
  tags?: string;
}